Update to libmicrohttpd v0.9.50

+++ ext/libmicrohttpd/mhd/src/microhttpd/basicauth.c
@@ -135,7 +135,7 @@ MHD_queue_basic_auth_fail_response (struct MHD_Connection *connection,
                        hlen, 
                        "Basic realm=\"%s\"", 
                        realm);
-  if (res > 0 && res < hlen)
+  if (res > 0 && (size_t)res < hlen)
     ret = MHD_add_response_header (response,
                                    MHD_HTTP_HEADER_WWW_AUTHENTICATE,
                                    header);
